checking for file existence..

+++ platform-modules/batch-models/src/test/scala/org/ekstep/analytics/job/TestPrecomputedViewsJob.scala
@@ -2,11 +2,15 @@ package org.ekstep.analytics.job
 
 import org.ekstep.analytics.model.BaseSpec
 import org.ekstep.analytics.framework.util.CommonUtil
+import java.io.File
 
 class TestPrecomputedViewsJob extends BaseSpec {
   
     it should "execute the PrecomputedViews successfully" in {
         PrecomputedViewsJob.main()
-        CommonUtil.deleteDirectory("/tmp/data-sets/");
+        val file = new File("/tmp/data-sets/")
+        if(file.exists()){
+            CommonUtil.deleteDirectory("/tmp/data-sets/"); 
+        }
     }
 }
